requirements.txt
eth-abi==2.1.1
eth-account==0.5.6
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
Flask==2.0.3
google-api-core==2.6.0
google-api-python-client==2.39.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.0
googleapis-common-protos==1.55.0
numpy==1.20.3
pipreqs==0.4.11
ptyprocess==0.7.0
urllib3==1.26.7
web3==5.25.0
google-cloud-core==2.2.3 
google-cloud-storage==2.2.1 
google-crc32c==1.3.0 
google-resumable-media==2.3.2
